requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
tensorboard>=2.13.0
torchmetrics>=1.0.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
PyQt6>=6.5.0
PyOpenGL>=3.1.6
pywavefront>=1.3.3
